+++ guicey-server-pages/src/main/java/ru/vyarus/guicey/gsp/app/util/PathUtils.java
@@ -14,7 +14,7 @@ import java.util.regex.Pattern;
 public final class PathUtils {
 
     public static final String SLASH = "/";
-    private static final Pattern PATH_DIRTY_SLASHES = Pattern.compile("\\s*/\\s*/+\\s*");
+    private static final Pattern PATH_DIRTY_SLASHES = Pattern.compile("\\s*/\\s*(/+\\s*)?");
     private static final CharMatcher TRIM_SLASH = CharMatcher.is('/');
     private static final CharMatcher TRIM_STAR = CharMatcher.is('*');
 
@@ -29,14 +29,14 @@ public final class PathUtils {
      * @return path combined with base path
      */
     public static String normalizePath(final String basePath, final String path) {
-        if (path == null || path.isEmpty()) {
-            return basePath;
-        }
-        return path(basePath, path);
+        // when no path to merge still apply same normalization rules for more predictable result
+        return path == null || path.isEmpty() ? path(basePath)
+                // empty base path will bring unexpected '/' prefix (due to join)
+                : (basePath == null || basePath.isEmpty() ? path(path) : path(basePath, path));
     }
 
     /**
-     * If parts may contain slashes (leading / trailing) they will be cleaned out, so the resulted
+     * If parts contain slashes (leading / trailing) they will be cleaned out, so the resulted
      * path will not contain double slashes.
      *
      * @param parts path parts
@@ -50,10 +50,11 @@ public final class PathUtils {
      * Cleanup duplicate slashes and replace backward slashes.
      *
      * @param path path to cleanup
-     * @return path with canonical slashed
+     * @return path with canonical slashes
      */
     public static String cleanUpPath(final String path) {
-        return PATH_DIRTY_SLASHES.matcher(path).replaceAll(SLASH).trim();
+        final String fixedBackslashes = path.replace('\\', '/');
+        return PATH_DIRTY_SLASHES.matcher(fixedBackslashes).replaceAll(SLASH).trim();
     }
 
     /**
@@ -99,10 +100,47 @@ public final class PathUtils {
     }
 
     /**
+     * Returned location path does not contain leading slash because its not needed for direct classpath resource
+     * loading.
+     *
      * @param cls class
      * @return class location path
      */
     public static String getPath(final Class cls) {
         return cls.getPackage().getName().replace(".", PathUtils.SLASH);
+    }
+
+    /**
+     * Normalization for sub section (sub folder) url path. Rules:
+     * <ul>
+     *     <li>Backslashes replaced with '/'</li>
+     *     <li>Url must not starts with '/'</li>
+     *     <li>Url must end with '/' (to prevent false sub-string matches when path used for matches)</li>
+     * </ul>
+     *
+     * @param path path
+     * @return normalized url
+     */
+    public static String normalizeRelativePath(final String path) {
+        final String cleanPath = cleanUpPath(path);
+        // do not apply end slash to empty path to not confuse with leading slash
+        // NOTE '/' will become '' - intentional!
+        return cleanPath.isEmpty() ? cleanPath : endSlash(trimSlashes(cleanPath));
+    }
+
+    /**
+     * Normalization for classpath resource path. Rules:
+     * <ul>
+     *     <li>Path use '/' as separator</li>
+     *     <li>Backslashes replaced with '/'</li>
+     *     <li>Path must not start with '/'</li>
+     *     <li>Path ends with '/'</li>
+     * </ul>
+     *
+     * @param path classpath path (with '.' or '/')
+     * @return normalized classpath path
+     */
+    public static String normalizeClasspathPath(final String path) {
+        return normalizeRelativePath(path.replace('.', '/'));
     }
 }