+++ guicey-server-pages/src/main/java/ru/vyarus/guicey/gsp/app/util/ResourceLookup.java
@@ -4,10 +4,10 @@ import com.google.common.base.CharMatcher;
 import com.google.common.base.MoreObjects;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import ru.vyarus.guicey.gsp.app.asset.AssetLookup;
 import ru.vyarus.guicey.gsp.views.template.TemplateNotFoundException;
 
 import java.util.Iterator;
-import java.util.List;
 
 /**
  * Utility used to lookup static resources in multiple locations. This is used for applications extensions
@@ -42,7 +42,7 @@ public final class ResourceLookup {
      * @param rootPaths classpath folders to search resource in
      * @return resource location path (first occurrence) or null if not found
      */
-    public static String lookup(final String path, final List<String> rootPaths) {
+    public static String lookup(final String path, final Iterable<String> rootPaths) {
 
         final String templatePath = CharMatcher.is('/').trimLeadingFrom(path);
         final Iterator<String> it = rootPaths.iterator();
@@ -58,18 +58,19 @@ public final class ResourceLookup {
     }
 
     /**
-     * Shortcut for {@link #lookup(String, List)} with fail in case of not found template.
+     * Shortcut for {@link #lookup(String, Iterable)} with fail in case of not found template.
      *
-     * @param path      static resource path
-     * @param rootPaths classpath folders to search resource in
+     * @param path   static resource path
+     * @param assets assets resolution object
      * @return resource location path (first occurrence)
      * @throws TemplateNotFoundException if template not found
      */
-    public static String lookupOrFail(final String path, final List<String> rootPaths)
+    public static String lookupOrFail(final String path, final AssetLookup assets)
             throws TemplateNotFoundException {
-        final String lookup = ResourceLookup.lookup(path, rootPaths);
+        final String lookup = assets.lookup(path);
         if (lookup == null) {
-            final String err = String.format("Template %s not found in locations: %s", path, rootPaths);
+            final String err = String.format(
+                    "Template %s not found in locations: %s", path, assets.getMatchingLocations(path));
             // logged here because exception most likely will be handled as 404 response
             LOGGER.info(err);
             throw new TemplateNotFoundException(err);